TamperMonkey scripts: make style fix work again.
Ultra-minimal effort with minimal testing because I hardly use the site
anyhow, it seems to work.

Amazing that someone still appears to work on the site at all. I thought
it would be dead by now. But I'm starting to suspect that these
atrocious CSS styles are part of the plot to make the site so annoying
to use that people will eventually abandon it and they can finally pull
the plug without too much protest.